+++ src/main/java/com/sparta/barointern/application/security/UserDetailsServiceImpl.java
@@ -1,9 +1,9 @@
-package com.sparta.barointern.infrastructure.security;
+package com.sparta.barointern.application.security;
 
 import com.sparta.barointern.domain.entity.User;
 import com.sparta.barointern.domain.repository.UserRepository;
-import com.sparta.barointern.exception.BaseException;
-import com.sparta.barointern.exception.Code;
+import com.sparta.barointern.infrastructure.exception.BaseException;
+import com.sparta.barointern.infrastructure.exception.Code;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
