+++ src/main/java/com/sparta/barointern/domain/entity/User.java
@@ -8,12 +8,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Getter
-@AllArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
 public class User {
-
-
 
     private String username;
     private String nickname;
@@ -22,22 +20,19 @@ public class User {
     private List<Role> roles= new ArrayList<>(List.of(Role.create(UserRole.USER)));
 
     public void grantAdmin() {
-        if (this.roles == null) { // 혹시 roles가 초기화 안됐을 수도 있으니 방어 코드
+        if (this.roles == null) {
             this.roles = new ArrayList<>();
         }
-        // 이미 ADMIN 역할이 있는지 확인하고 없으면 추가
         if (!this.roles.contains(UserRole.ADMIN)) {
             this.roles.set(0,Role.create(UserRole.ADMIN));
         }
     }
-
 
     public static User create(String username, String password, String nickname) {
         return User.builder()
                 .username(username)
                 .nickname(nickname)
                 .password(password)
-//                .roles(roles)
                 .build();
     }
