+++ src/main/java/com/sparta/barointern/infrastructure/config/WebSecurityConfig.java
@@ -3,19 +3,17 @@ package com.sparta.barointern.infrastructure.config;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.barointern.common.ErrorResponse;
 import com.sparta.barointern.common.Process;
-import com.sparta.barointern.domain.enums.UserRole;
-import com.sparta.barointern.exception.Code;
-import com.sparta.barointern.exception.CustomAuthenticationEntryPoint;
+import com.sparta.barointern.infrastructure.exception.Code;
+import com.sparta.barointern.infrastructure.exception.CustomAuthenticationEntryPoint;
 import com.sparta.barointern.infrastructure.jwt.JwtAuthenticationFilter;
 import com.sparta.barointern.infrastructure.jwt.JwtAuthorizationFilter;
 import com.sparta.barointern.infrastructure.jwt.JwtUtil;
-import com.sparta.barointern.infrastructure.security.UserDetailsServiceImpl;
+import com.sparta.barointern.application.security.UserDetailsServiceImpl;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
