+++ src/main/java/com/sparta/barointern/infrastructure/exception/GlobalExceptionHandler.java
@@ -1,12 +1,12 @@
-package com.sparta.barointern.exception;
+package com.sparta.barointern.infrastructure.exception;
 
-import com.sparta.barointern.common.ApiResponse;
 import com.sparta.barointern.common.ErrorResponse;
 import com.sparta.barointern.common.Process;
+import com.sparta.barointern.domain.entity.Role;
+import com.sparta.barointern.domain.entity.User;
 import io.swagger.v3.oas.annotations.Hidden;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
@@ -18,8 +18,8 @@ public class GlobalExceptionHandler {
     public ResponseEntity<ErrorResponse> handlerBaseExceptionException(BaseException e) {
         Code code = e.getCode();
         Process error = Process.from(code);
-
         return ResponseEntity.status(code.getStatus()).body(ErrorResponse.from(error));
     }
+
 
 }
