+++ src/main/java/com/sparta/barointern/infrastructure/jwt/JwtAuthorizationFilter.java
@@ -3,8 +3,8 @@ package com.sparta.barointern.infrastructure.jwt;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.barointern.common.ErrorResponse;
 import com.sparta.barointern.common.Process;
-import com.sparta.barointern.exception.Code;
-import com.sparta.barointern.infrastructure.security.UserDetailsServiceImpl;
+import com.sparta.barointern.infrastructure.exception.Code;
+import com.sparta.barointern.application.security.UserDetailsServiceImpl;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
